refactor: extrait les emails transactionnels de formulaire.service.ts
Candidat architecture (formulaire.service.ts, 5 responsabilites
melangees) : extraction ciblee des 2 fonctions d'envoi d'email
(sendDelegationMailToCFA, sendMailNouvelleOffre) + leur dependance
privee getJobOrigin vers un nouveau fichier
formulaire-notifications.service.ts.

Perimetre volontairement limite pour eviter tout conflit avec la PR
#5078 (MVP offres 100% libres, toujours ouverte) : verifie ligne par
ligne que #5078 ne touche ni ces fonctions ni leur voisinage. Les
fonctions de conversion recruiter<->jobsPartners (jobCreateToJobsPartner,
jobPartnersToRecruiter) restent en place car #5078 modifie directement
jobCreateToJobsPartner - a traiter apres son merge.

Note en passant : createJobDelegations contient encore un envoi
d'email inline (template mail-mer-confirmation) couple a l'etat de
sa boucle de delegation - candidat a une extraction future, non traite
ici pour rester dans le perimetre valide.
