+++ server/src/jobs/recruiters/update-siret-infos-in-error-job.ts
@@ -13,7 +13,8 @@ import { getDbCollection } from "@/common/utils/mongodb-utils"
 import { sentryCaptureException } from "@/common/utils/sentry-utils"
 import { notifyToSlack } from "@/common/utils/slack-utils"
 import { getEntrepriseDataFromSiret } from "@/services/etablissement.service"
-import { archiveFormulaire, sendMailNouvelleOffre } from "@/services/formulaire.service"
+import { archiveFormulaire } from "@/services/formulaire.service"
+import { sendMailNouvelleOffre } from "@/services/formulaire-notifications.service"
 import { upsertEntrepriseData } from "@/services/organization.service"
 import { sendDeactivatedRecruteurMail } from "@/services/role-management.service"
 import { setEntrepriseInError } from "@/services/user-recruteur.service"
